% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{compute_stat_wrangling}
\alias{compute_stat_wrangling}
\title{Scalar method for sf}
\usage{
compute_stat_wrangling(
  baseline_values,
  focal_value,
  stat_wrangling = c("deviation", "sd_deviation", "count_above", "count_below"),
  baseline_fun = function(x) mean(x, na.rm = TRUE)
)
}
\description{
Scalar method for sf
}
