% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{compute_stat_wrangling.SpatRaster}
\alias{compute_stat_wrangling.SpatRaster}
\title{Raster method for SpatRaster}
\usage{
\method{compute_stat_wrangling}{SpatRaster}(
  baseline_values,
  focal_value,
  stat_wrangling = c("deviation", "sd_deviation", "count_above", "count_below"),
  baseline_fun = function(x) mean(x, na.rm = TRUE)
)
}
\description{
Raster method for SpatRaster
}
