% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxc-package.R
\docType{package}
\name{gxc}
\alias{gxc-package}
\alias{gxc}
\title{gxc: Easy Access to Earth Observation Data}
\description{
This package is part of the project "GESIS meets Copernicus".
It supports users of Earth observation data by offering easy
access and processing of data from Copernicus services and other
Earth observation sources.
}
\details{
Version: 0.1.0

This package provides six core functions to link your input data
with EO indicators. Depending on your data input format
(points, polygons, or raster) and the desired temporal resolution
of the required indicators (daily or monthly), you can choose from the
list of functions below:

For Spatial Points (\code{point_link_})
\itemize{
\item \code{point_link_daily}
\item \code{point_link_monthly}
}

For Polygons (\code{poly_link_})
\itemize{
\item \code{poly_link_daily}
\item \code{poly_link_monthly}
}

For Gridded Data (\code{grid_link_})
\itemize{
\item \code{grid_link_daily}
\item \code{grid_link_monthly}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/denabel/gxc}
  \item Report bugs at \url{https://github.com/denabel/gxc/issues}
}

}
\author{
Dennis Abel \email{dennis.abel@gesis.org} and
Stefan Jünger \email{stefan.juenger@gesis.org}
}
