% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{gxc_indicators}
\alias{gxc_indicators}
\title{List all available indicators}
\usage{
gxc_indicators()
}
\value{
A dataframe with columns \code{catalogue}, \code{source}, and \code{indicator}.
}
\description{
Returns a dataframe of all indicators available in gxc,
grouped by catalogue.
}
\examples{
gxc_indicators()
}
