% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_keys.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API keys}
\usage{
set_api_key(service = "ecmwfr", user = NULL)
}
\arguments{
\item{service}{Service for which an API key should be stored. Currently,
only \code{ecmwfr} is supported, which is used in \code{\link{link_daily}} and
\code{\link{link_monthly}}.}

\item{user}{Optionally, a user name if multiple keys are set for a single
service. Not necessary most of the time.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Convenience function to set all kinds of API keys needed for the \code{gxc}
linking functions.
}
\examples{
\dontrun{
set_api_key("ecmwfr")
}
}
