% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_dwd_modules.R
\name{specify_days}
\alias{specify_days}
\title{Specify Days Based on Input Date and Criteria}
\usage{
specify_days(date, days_back, months)
}
\arguments{
\item{date}{A date object representing the reference date from which to
calculate the days.}

\item{days_back}{Optional integer specifying how many days back to include
from the reference date.}

\item{months}{Optional character vector specifying the months to consider
(e.g., "2023-01" for January 2023). If the month string starts with "P", it
indicates a previous year.}
}
\value{
A vector of date objects representing the specified days.
}
\description{
This function generates a sequence of days based on the provided date,
a specified number of days to go back, and/or specific months.
It allows for flexible selection of days either by a range of days back
from the given date or by specifying certain months.
}
\details{
The function checks for \code{days_back} and \code{months} parameters:
\itemize{
\item If both are NULL, it returns the original date.
\item If \code{days_back} is specified, it generates a sequence of days going back
from the given date.
\item If \code{months} is specified, it calculates the start and end dates for the
specified months and returns all days within that range.
}
}
\examples{
# Example 1: Specify days going back 5 days from a given date
specify_days(as.Date("2024-01-15"), days_back = 5, months = NULL)

# Example 2: Specify days for January and February of 2024
specify_days(as.Date("2024-01-15"), days_back = NULL, months = c(1, 2))

# Example 3: Specify days without additional criteria, returning the original 
date
specify_days(as.Date("2024-01-15"), days_back = NULL, months = NULL)

}
